+++ src/main/scala/org/galaxio/gatling/feeders/RandomJurITNFeeder.scala
@@ -1,7 +1,7 @@
 package org.galaxio.gatling.feeders
 
 import io.gatling.core.feeder.Feeder
-import org.galaxio.gatling.utils.RandomDataGenerators
+import org.galaxio.gatling.feeders.faker.GovIdGenerators
 
 @deprecated("Use org.galaxio.gatling.feeders.faker.Faker.ru.inn.company with GeneratedFeeder instead", "faker-api")
 object RandomJurITNFeeder {
@@ -16,6 +16,6 @@ object RandomJurITNFeeder {
     *   random string ITN of the juridical person feeder
     */
   def apply(paramName: String): Feeder[String] =
-    feeder[String](paramName)(RandomDataGenerators.randomJurITN())
+    feeder[String](paramName)(GovIdGenerators.jurITN())
 
 }
