+++ src/main/scala/org/galaxio/gatling/feeders/RandomKPPFeeder.scala
@@ -1,7 +1,7 @@
 package org.galaxio.gatling.feeders
 
 import io.gatling.core.feeder.Feeder
-import org.galaxio.gatling.utils.RandomDataGenerators
+import org.galaxio.gatling.feeders.faker.GovIdGenerators
 
 @deprecated("Use org.galaxio.gatling.feeders.faker.Faker.ru.kpp with GeneratedFeeder instead", "faker-api")
 object RandomKPPFeeder {
@@ -16,6 +16,6 @@ object RandomKPPFeeder {
     *   random string KPP feeder
     */
   def apply(paramName: String): Feeder[String] =
-    feeder[String](paramName)(RandomDataGenerators.randomKPP())
+    feeder[String](paramName)(GovIdGenerators.kpp())
 
 }
