+++ src/main/scala/org/galaxio/gatling/utils/RandomDataGenerators.scala
@@ -6,8 +6,8 @@ import java.time.format.DateTimeFormatter
 import java.time.temporal.TemporalUnit
 import java.time.{Instant, LocalDateTime, ZoneId}
 import java.util.UUID
-import scala.annotation.tailrec
 import scala.util.Random
+import org.galaxio.gatling.feeders.faker.GovIdGenerators
 
 object RandomDataGenerators {
 
@@ -159,184 +159,29 @@ object RandomDataGenerators {
       max
   }
 
-  /** Selects a random element from the provided list of integers. If the list is empty, a random integer of the specified
-    * length is generated and returned.
-    *
-    * @param items
-    *   a list of integers to select a random element from; can be empty
-    * @param intLength
-    *   the number of digits for generating a random integer if the list is empty
-    * @return
-    *   a random integer from the list if it is non-empty, or a randomly generated integer of the specified length if the list
-    *   is empty
-    */
-  private def getRandomElement(items: List[Int], intLength: Int): Int = items match {
-    case Nil => randomValue(intLength)
-    case _   => items(randomValue(items.length))
-  }
+  @deprecated("Use Faker.finance.pan (with GeneratedFeeder) instead", "faker-api")
+  def randomPAN(bins: String*): String = GovIdGenerators.pan(bins: _*)
 
-  /** Selects a random element from the given list of strings or generates a random string if the list is empty.
-    *
-    * @param items
-    *   the list of strings to select a random element from; can be empty
-    * @param stringLength
-    *   the length of the random string to generate if the list is empty; must be greater than 0
-    * @return
-    *   a randomly selected string from the list or a randomly generated string of the specified length
-    * @throws IllegalArgumentException
-    *   if `stringLength` is less than or equal to 0
-    */
-  private def getRandomElement(items: List[String], stringLength: Int): String = items match {
-    case Nil => lettersString(stringLength)
-    case _   => items(randomValue(items.length))
-  }
+  @deprecated("Use Faker.ru.ogrn (with GeneratedFeeder) instead", "faker-api")
+  def randomOGRN(): String = GovIdGenerators.ogrn()
 
-  // Hoisted check-digit weight tables (FTS algorithm) — read-only, shared across virtual users, so
-  // allocated once at object init instead of on every generated id / every recursion step.
-  private val NatInnFactors1: Array[Int] = Array(7, 2, 4, 10, 3, 5, 9, 4, 6, 8)
-  private val NatInnFactors2: Array[Int] = Array(3, 7, 2, 4, 10, 3, 5, 9, 4, 6, 8)
-  private val JurInnFactors: Array[Int]  = Array(2, 4, 10, 3, 5, 9, 4, 6, 8)
+  @deprecated("Use Faker.ru.ogrnip (with GeneratedFeeder) instead", "faker-api")
+  def randomPSRNSP(): String = GovIdGenerators.ogrnip()
 
-  /** Left-pad a non-negative int with zeros to `width` (cheap replacement for `String.format("%0Nd")`). */
-  private def zeroPad(value: Int, width: Int): String = {
-    val s = value.toString
-    if (s.length >= width) s else "0" * (width - s.length) + s
-  }
+  @deprecated("Use Faker.ru.kpp (with GeneratedFeeder) instead", "faker-api")
+  def randomKPP(): String = GovIdGenerators.kpp()
 
-  private def digitsToString(digits: Array[Int]): String = {
-    val sb = new java.lang.StringBuilder(digits.length)
-    var i  = 0
-    while (i < digits.length) {
-      sb.append(('0' + digits(i)).toChar)
-      i += 1
-    }
-    sb.toString
-  }
+  @deprecated("Use Faker.ru.inn.person (with GeneratedFeeder) instead", "faker-api")
+  def randomNatITN(): String = GovIdGenerators.natITN()
 
-  def randomPAN(bins: String*): String = {
-    val idNum: String = digitString(9)
-    val body: String  = bins.toList match {
-      case Nil => s"${digitString(6)}$idNum"
-      case bs  => s"${getRandomElement(bs, 6)}$idNum"
-    }
+  @deprecated("Use Faker.ru.inn.company (with GeneratedFeeder) instead", "faker-api")
+  def randomJurITN(): String = GovIdGenerators.jurITN()
 
-    // Luhn: double every second digit from the left of the body; the check digit makes the total ≡ 0 (mod 10).
-    var sum        = 0
-    var i          = 0
-    while (i < body.length) {
-      val d = body.charAt(i).asDigit
-      sum += (if (i % 2 == 0) { val doubled = d * 2; if (doubled > 9) doubled - 9 else doubled }
-              else d)
-      i += 1
-    }
-    val controlNum = (10 - sum % 10) % 10
-    s"$body$controlNum"
-  }
+  @deprecated("Use Faker.ru.snils (with GeneratedFeeder) instead", "faker-api")
+  def randomSNILS(): String = GovIdGenerators.snils()
 
-  def randomOGRN(): String = {
-    val indicatorOGRN: Int   = getRandomElement(List(1, 5), 1)
-    val year: String         = zeroPad(randomValue(2, 21), 2)
-    val ruSubjectNum: String = zeroPad(randomValue(1, 90), 2)
-    val idNum: String        = digitString(7)
-    val result: String       = s"$indicatorOGRN$year$ruSubjectNum$idNum"
-    val rem: Long            = result.toLong % 11
-
-    if (rem == 10) s"${result}0" else s"$result$rem"
-  }
-
-  def randomPSRNSP(): String = {
-    val indicatorPSRNSP: Int = 3
-    val year: String         = zeroPad(randomValue(2, 21), 2)
-    val ruSubjectNum: String = zeroPad(randomValue(1, 90), 2)
-    val idNum: String        = digitString(9)
-    val result: String       = s"$indicatorPSRNSP$year$ruSubjectNum$idNum"
-    // (first 14 digits mod 13) mod 10 — always 0..9, so no mod-11-style "== 10" special case is needed.
-    val rem: Long            = result.toLong % 13 % 10
-
-    s"$result$rem"
-  }
-
-  def randomKPP(): String = {
-    val revenueServiceCode: String = zeroPad(randomValue(1, 10000), 4)
-    val reasonForReg: String       = zeroPad(randomValue(1, 100), 2)
-    val idNum: String              = zeroPad(randomValue(1, 1000), 3)
-
-    s"$revenueServiceCode$reasonForReg$idNum"
-  }
-
-  // Natural person (физлицо) ITN = 12 digits with TWO control digits.
-  // NOTE: corrected — earlier releases swapped this with randomJurITN (emitted a 10-digit legal-entity value).
-  def randomNatITN(): String = {
-    val d    = new Array[Int](12)
-    var i    = 0
-    while (i < 10) {
-      d(i) = randomValue(0, 10)
-      i += 1
-    }
-    var sum1 = 0
-    i = 0
-    while (i < 10) {
-      sum1 += d(i) * NatInnFactors1(i)
-      i += 1
-    }
-    d(10) = sum1 % 11 % 10
-    var sum2 = 0
-    i = 0
-    while (i < 11) {
-      sum2 += d(i) * NatInnFactors2(i)
-      i += 1
-    }
-    d(11) = sum2 % 11 % 10
-    digitsToString(d)
-  }
-
-  // Legal entity (юрлицо) ITN = 10 digits with ONE control digit.
-  // NOTE: corrected — earlier releases swapped this with randomNatITN (emitted a 12-digit natural-person value).
-  def randomJurITN(): String = {
-    val d   = new Array[Int](10)
-    var i   = 0
-    while (i < 9) {
-      d(i) = randomValue(0, 10)
-      i += 1
-    }
-    var sum = 0
-    i = 0
-    while (i < 9) {
-      sum += d(i) * JurInnFactors(i)
-      i += 1
-    }
-    d(9) = sum % 11 % 10
-    digitsToString(d)
-  }
-
-  def randomSNILS(): String = {
-    val d       = new Array[Int](9)
-    var i       = 0
-    while (i < 9) {
-      d(i) = randomValue(0, 10)
-      i += 1
-    }
-    var sum     = 0
-    i = 0
-    while (i < 9) {
-      sum += d(i) * (9 - i) // first digit ×9 .. ninth digit ×1
-      i += 1
-    }
-    // control = sum mod 101; 100 is written as "00" (101 cannot occur after the mod).
-    val control = sum % 101 match {
-      case 100 => 0
-      case c   => c
-    }
-    s"${digitsToString(d)}${zeroPad(control, 2)}"
-  }
-
-  def randomRusPassport(): String = {
-    val ruSubjectNum: String = String.format("%02d", randomValue(1, 90))
-    val year: String         = String.format("%02d", randomValue(0, 21))
-    val idNum: String        = digitString(6)
-
-    s"""$ruSubjectNum$year$idNum"""
-  }
+  @deprecated("Use Faker.passport.ru (with GeneratedFeeder) instead", "faker-api")
+  def randomRusPassport(): String = GovIdGenerators.rusPassport()
 
   private def validateDelta(positiveDelta: Int, negativeDelta: Int): Unit = {
     require(
